requirements.txt
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
flower==2.0.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanize==4.12.3
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
kombu==5.5.4
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
orjson==3.11.1
packaging==25.0
passlib==1.7.4
prometheus-client==0.22.1
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
redis==6.2.0
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
sentry-sdk==2.33.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.47.2
tornado==6.5.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
